requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
hyperon>=0.1.0
flask>=2.0.0
